rename error boundary and add JSDocs
