+++ src/platform/packages/shared/kbn-typed-react-router-config/index.ts
@@ -20,3 +20,4 @@ export * from './src/use_params';
 export * from './src/use_router';
 export * from './src/use_route_path';
 export * from './src/breadcrumbs';
+export * from './src/route_self_heal_error_boundary';
